010 Generic Take Action (11:23) / Chapter 004 Actions & UI:
Tried adding a more generic way of inserting INPUT parameters to the Generic function call to 'TakeAction'(...)... the goal is to use it as:    TakeAction(baseParameters)   for every kind of Action. BaseParameters is an Abstract Class that is reimplemented by means of Inheritance as, for instance:  'public MoveBaseParameters : BaseParameters'  for the case of MOveAction Class (which will use it).

This is an experimental idea, based on the best practices for Architecture Patters for AAA Games.