014 Turn System (15:57) / Chapter 004 Actions & UI:
1- (Continued with...) Created a **Turn System** that will ***manage the turn*** and ***reset*** the Unit’s *Action Points*

* To end this Lecture, I added the feature of RESETTING A TURN, which also means: resetting the ActionPoints.
For this task, I added multiple (4) more Delegates (EVENT Handlers & Listeners) to listen to any Events associated with any changes in the _actionPoints  variable/Attribute of the Unit.cs's  Class (and also some Delegates are related to any changes in the Current Turn... are triggered when going to the NEXT TURN).

THE END (of this Lecture and also CHAPTER 04).

NEXT:  Setting the Enemies:  Animations, A.I. & MORE!