005 Shooting Bullet Visual (20:51) / Chapter 005 Enemies & Combat:
1- The **Trail Renderer** and **Particle System** are been used to make a nice ***Bullet visual*** and ***explosion***

Details: This is an intermediate Commit. Up to this Point:

* Implemented the feature for:
The BulletProjectile.prefab was (previously) been shot towards the Feet of the Enemy. That  is fixed in this commit, by calculating the direction of the Movement, setting the y-value as a Constant (pointing towards the Geometric Center of the Target), right from the beginning of the Bullet Animation movement.