* I added the ***GridSystemVisual*** to ***run after** the **Default Time*** since you want the **visuals** to **run after** everything else.
  NOTE:

* This change was inspired because there is a discussion on the Course Forums (https://community.gamedev.tv/t/race-condition-with-the-unit-referenced-in-the-editor-on-unitactionsystem/205735); about a Race condition with the unit referenced in the editor on UnitActionSystem: what happens is that, in the words of the Teacher Assistant (although he rather calls himself: "Master Problem Solver"), named Bixarrio:

"We have defined a default selected unit on the UnitActionSystem but there is a race condition (at least on my machine) where the GridSystemVisual.Start() runs before the Unit.Start(). This then starts the game with my default unit - which is not at position 0,0 - to be displayed incorrectly because the Unit has not retrieved its grid position yet and returns the default 0,0. Once I have selected a different unit, everything is fine. It’s just visually not displaying the correct things at start. (...)".

* I changed some (Action Points) Maximum Values in the Player Units (GameObjects in the Scene...) to allow the Player to practice the MoveAction freely around the whole Game Board (and thus, the Pathfinding).