* I continued writing the "More Complex ENEMY A.I. Algorithm".
* Now it can Move towards its target in a multi-step approach (the thanks to the routine: **GetBestEnemyAIActionDataForMovingSimplyTowardsAGoal**):
The ENEMY A.I. can now walk towards its "Enemy", and stop its movement in a "Safe Area" where it cannot be shot by its enemy, so the ActionPoints can be reestablished when changing Turns, in the Game.
It is not finished yet: There are more details to add and some validations (related to the random numbers I want to use...). The "walking towards its Enemy (and waiting in a Safe Area until he has enough ActionPoints for a ShootAction)" part is done and Tested out.

* It is in a "early Alpha stage", in which there are some more Functions yet to be written, so by now the functionality is rough, but good enough for a first version (v-1.0). The actual final version of the algorithm will be written during the next days.

* All functionality is working and running, there are no compilation / build errors.