* I almost finish writing the "More Complex ENEMY A.I. Algorithm".
* The only things TODO are:

a) Adding one more validation.
b) Checking the code again, for cleanup.

* Also I will do a Refactor in the future, for perfecting and laking possible to extend the functionality into more general cases.

The ENEMY A.I. can now walk towards its "Enemy", and stop its movement in a "Safe Area" where it cannot be shot by its enemy, so the ActionPoints can be reestablished when changing Turns, in the Game.

* It is in a "early Beta stage", and I am Beta-Testing it (so far, very satisfactorily).

* All functionality is working and running, there are no compilation / build errors.