* Fixed Bug:  Spawning an "Human Player (Prefab) Ragdoll"  when an UnitEnemy_v2.prefab was killed:
Real Problem:

The UnitEnemy_v2.prefab (when spawned into the scene, as a GameObject...) was referencing another (Component) RagdollSpawner (and Ragdoll Prefab, called PlayerRagdoll instead of the one called EnemyRagdoll), with another set of Bones. It happened because the Base for that Component was just copied (mindlessly) from the UnitPlayer Prefab.

Solution:

Simply to reference in the PlayerEnemy's Component (called:  RagdollSpawners.cs) the correct set of 13 Bones that belong to an "EnemyRagdoll".