* New VFX Materialize for 3D Characters Effect: Additional features:
* Tutorial Video: [Unity VFX Graph - Healing Spell Effect Tutorial - YouTube](https://www.youtube.com/watch?v=ZCtEQrEOx_M)

1- Worked to make better "Particles". We could call this work: "Perfecting the Particles..."

Added:

a) Perfected the 'Texture 2D' used for the Small Laser Particle beams (they are larger now).

b) A waving effect to the thickness / width of the Main Cylinder (For that I used a node called:  Multiply Scale.XYZ Over Life, with a funny Animation Curve, similar (but not exactly) to: Sin(x) over the Line:  y=Sin(x) ).

===================

Todo in the future:

1- Adjusting the Colors of the Rings vs. Particles in the VFX.

2- Adjusting the Time (Duration) and Speed of the Rings vs. Particles vs. the Character's Materializing Effect in it's own 'Shader Graph' effect.
Note: In general terms, I want the VFX to be 20% quicker / faster.