* New VFX Materialize for 3D Characters Effect: Additional features:
Worked to make better "Particles". We could call this work: "Perfecting the Particles..."

Added:

1- Adjusted the Time (Duration) and Speed of the Rings vs. Particles vs. the Character's Materializing Effect in it's own 'Shader Graph' effect.
Note: In general terms, I wanted the VFX to be 20% quicker / faster. It is Done now.

================

Todo in the next commits:

1- Adjusting the Colors of the Rings vs. Particles in the VFX.

2- Re-adjust, (again), the Time (Duration) and Speed of the Rings vs. Particles vs. the Character's Materializing Effect in it's own 'Shader Graph' effect:

Why?

I just need to do this when changing the 3D Character to another (i.e.: the Polygon Prototype one...).

3- Set (enable) in the Main VFX Code (in the 3D Character's Inspector's GameObject) the option:  Delete this VFX when done (maybe after a couple of seconds, after it apparently ended).

4- Save the Prefab for using it in the scene, and link it to every 3D Character's Spawn Method execution.