* New VFX Materialize for 3D Characters Effect: Additional features:
Done

1- Optimization Improvement: Fixed error in: New feature: Reverting the Materials to Default ones (for keeping the scene optimized). in the Main VFX Code.
The Error was that the Materials were not swapping, because the reference to the SkinnedMeshRenderer's Array was not correctly (we cannot pass the SKinnedMeshRenderer Materials[] array BY REFERENCE correctly: to a function to set its materials. It just does not work that way. The workaround: to get a new fresh reference to SkinnedMeshRenderer in the function where I will Assign the Materials and make the changes).

That implied some other tasks on that VFX's Controller Script: added new attributes and methods for validation of the Materials, etc.

=======Todo==========

1- Add: David's Recommendations: (Perfecting the VFX). To the Top and Bottom's Ring Texture.

2- Clean the scene: Remove all the Experiments, and leave it as the game will be.